+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_ecma.dart
@@ -123,6 +123,10 @@ class _FormatterStandaloneECMA extends FormatterStandaloneImpl {
   @override
   String formatInternal(DateTime datetime) =>
       dateTimeFormat.format(datetime.js);
+
+  @override
+  String formatRangeInternal(DateTime start, DateTime end) =>
+      dateTimeFormat.formatRange(start.js, end.js);
 }
 
 class _FormatterECMA extends FormatterImpl {
@@ -141,6 +145,10 @@ class _FormatterECMA extends FormatterImpl {
   @override
   String formatInternal(DateTime datetime) =>
       dateTimeFormat.format(datetime.js);
+
+  @override
+  String formatRangeInternal(DateTime start, DateTime end) =>
+      dateTimeFormat.formatRange(start.js, end.js);
 
   @override
   ZonedDateTimeFormatter withTimeZoneLong() =>
@@ -530,6 +538,7 @@ extension type _DateTimeFormat._(JSObject _) implements JSObject {
     _DateTimeJSOptions options,
   ]);
   external String format(Date num);
+  external String formatRange(Date startDate, Date endDate);
 
   external static JSArray<JSString> supportedLocalesOf(JSArray listOfLocales);